requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-appconf==1.0.4
django-filter==2.4.0
django-imagekit==4.0.2
djangorestframework==3.12.4
idna==3.2
iniconfig==1.1.1
packaging==21.0
pilkit==2.0
Pillow==8.3.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-mock==3.6.1
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
